Create lyrics.py
This is a basic app that prints out lyrics when the song and song artist are mentioned. If you don't know any songs by the artist, the app provides you with top 10 songs from the artist. It takes a few minutes to fetch the songs. It is an API based app.